feat :: 뷰 모델 역할 분리
