+++ lib/data/view_models/item_rental_view_model.dart
@@ -1,51 +1,56 @@
+import 'package:danuri_flutter/core/provider/item_id_provider.dart';
+import 'package:danuri_flutter/core/provider/rental_id_provider.dart';
 import 'package:danuri_flutter/data/data_sources/other/item_rental_data_source.dart';
-import 'package:danuri_flutter/data/data_sources/other/space_data_source.dart';
+import 'package:danuri_flutter/data/models/other/common/request/usage_id_request.dart';
 import 'package:danuri_flutter/data/models/other/rental_item/request/rental_item_request.dart';
-import 'package:danuri_flutter/data/models/other/rental_item/response/available_items_response.dart';
-import 'package:danuri_flutter/data/models/other/rental_item/response/rented_item_response.dart';
-import 'package:danuri_flutter/data/models/other/space/response/space_usage_response.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
 class ItemRentalViewModel {
-  final ItemRentalDataSource _itemDataSource = ItemRentalDataSource();
-  final SpaceDataSource _spaceDataSource = SpaceDataSource();
-
-  RentedItemResponse? _rentedItem;
-  RentedItemResponse? get rentedItem => _rentedItem;
-
-  List<ItemAvailableRental>? _itemAvailableRental;
-  List<ItemAvailableRental>? get itemAvailableRental => _itemAvailableRental;
-
-  SpaceUsageResponse? _spaceUsage;
-  SpaceUsageResponse? get spaceUsage => _spaceUsage;
+  final _dataSource = ItemRentalDataSource();
 
   bool? _error;
   bool? get error => _error;
 
-  Future<void> itemRental(String usageId, String itemId, int quantity) async {
+  void reset() {
+    _error = false;
+  }
+
+  Future<void> itemRental(
+      BuildContext context, int quantity, String usageId) async {
     try {
-      _rentedItem = await _itemDataSource.itemRental(
-        usageId,
-        RentalItemRequest(itemId: itemId, quantity: quantity),
+      final String itemId = context.read<ItemIdProvider>().itemId;
+      // await SpaceDataSource().getUsageSpace().then(
+      //   (usageSpace) async {
+      //     if (usageSpace.spaceUsageInfo == null) {
+      //       _error = true;
+      //     } else {
+      await _dataSource
+          .itemRental(
+        RentalItemRequest(
+          itemId: itemId,
+          quantity: quantity,
+          usageId: usageId,
+        ),
+      )
+          .then(
+        (rentedItem) {
+          context.read<RentalIdProvider>().addRentalId(rentedItem.id);
+        },
       );
       _error = false;
+      //     }
+      //   },
+      // );
     } on DioException catch (_) {
       _error = true;
     }
   }
 
-  Future<void> getItemAvailableRent() async {
+  Future<void> returnItem(String usageId) async {
     try {
-      _itemAvailableRental = await _itemDataSource.getItemAvailableRental();
-      _error = false;
-    } on DioException catch (_) {
-      _error = true;
-    }
-  }
-
-  Future<void> getUsageSpace() async {
-    try {
-      _spaceUsage = await _spaceDataSource.getUsageSpace();
+      await _dataSource.returnItem(UsageIdRequest(usageId: usageId));
       _error = false;
     } on DioException catch (_) {
       _error = true;
